pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "file-stats-mcp"
version = "2.3.0"
description = "一个基于MCP的文件统计与分析工具，提供全面的文件系统分析功能"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["mcp", "file-system", "statistics", "analysis", "tools"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Filesystems",
    "Topic :: Utilities",
]
dependencies = [
    "fastmcp>=0.4.1",
]

[project.urls]
Homepage = "https://github.com/yourusername/file-stats-mcp"
Repository = "https://github.com/yourusername/file-stats-mcp.git"
Issues = "https://github.com/yourusername/file-stats-mcp/issues"

[project.scripts]
file-stats-mcp = "mcp_server:main"

[tool.setuptools]
py-modules = ["mcp_server"]